fix: ForegroundService
Подключил пример с ForegroundService и настроил согласно задания.
